move objects to own component
