+++ src/data/AircraftData.h
@@ -28,7 +28,7 @@
 #include <utility>
 #include <vector>
 
-#include "object/Aircraft.h"
+#include "../object/Aircraft.h"
 #include "processor/AircraftProcessor.h"
 #include "Data.h"
