+++ src/feed/client/Client.h
@@ -28,14 +28,17 @@
 
 #include "../../Parameters.h"
 
+/// @namespace feed
 namespace feed
 {
 class Feed;
 
+/// @namespace client
 namespace client
 {
-/// Wait for (re-)connect timeout
 #ifdef CLIENT_CONNECT_WAIT_TIMEVAL
+/// @def C_CON_WAIT_TIMEVAL
+/// /// Wait for (re-)connect timeout
 #define C_CON_WAIT_TIMEVAL CLIENT_CONNECT_WAIT_TIMEVAL
 #else
 #define C_CON_WAIT_TIMEVAL 120
@@ -43,20 +46,16 @@ namespace client
 
 /**
  * @class Client
- * @brief Incomplete base-class representing a TCP client.
+ * @brief Base class representing a TCP client.
  * @note A Client is unique and only movable.
  */
 class Client
 {
 public:
-    /**
-     * Non-copyable
-     */
+    /// No copy construction
     Client(const Client&) = delete;
 
-    /**
-     * Not assignable
-     */
+    /// No copy assignment
     Client& operator=(const Client&) = delete;
 
     /**
@@ -69,7 +68,7 @@ public:
      * @fn run
      * @brief Run the Client.
      * @note Returns after all queued handlers have returned.
-     * @param r_sigset The signal set reference to register handler
+     * @param rSigset The signal set reference to register handler
      */
     void run(boost::asio::signal_set& rSigset);
 
@@ -83,10 +82,10 @@ protected:
     /**
      * @fn Client
      * @brief Constructor
-     * @param cr_host  The hostname
-     * @param cr_port  The port
-     * @param cr_comp  The component name
-     * @param r_feed   The handler Feed reference
+     * @param crHost       The hostname
+     * @param crPort       The port
+     * @param crComponent  The component name
+     * @param rFeed        The handler Feed reference
      */
     Client(const std::string& crHost, const std::string& crPort,
            const std::string& crComponent, feed::Feed& rFeed);
@@ -106,22 +105,21 @@ protected:
     /**
      * @fn connect
      * @brief Connect to host.
-     * @note To be implemented.
      */
     virtual void connect() = 0;
 
     /**
      * @fn handleTimedConnect
      * @brief Handler for timedConnect.
-     * @param cr_ec The error code
+     * @param crError The error code
      */
     void handleTimedConnect(const boost::system::error_code& crError) noexcept;
 
     /**
      * @fn handleRead
      * @brief Handler for read.
-     * @param cr_ec The error code
-     * @param s     The sent bytes
+     * @param crError The error code
+     * @param vBytes  The sent bytes
      */
     void handleRead(const boost::system::error_code& crError,
                     std::size_t vBytes) noexcept;
@@ -129,9 +127,8 @@ protected:
     /**
      * @fn handleResolve
      * @brief Handler for resolve host.
-     * @note Internally used in connect. To be implemented.
-     * @param cr_ec The error code
-     * @param it    The resolve iterator
+     * @param crError     The error code
+     * @param vResloverIt The resolve iterator
      */
     virtual void
     handleResolve(const boost::system::error_code& crError,
@@ -141,46 +138,55 @@ protected:
     /**
      * @fn handleConnect
      * @brief Handler for connect.
-     * @note To be implemented.
-     * @param cr_ec The error code
-     * @param it    The resolve iterator
+     * @param crError     The error code
+     * @param vResolverIt The resolve iterator
      */
     virtual void
     handleConnect(const boost::system::error_code& crError,
                   boost::asio::ip::tcp::resolver::iterator vResolverIt) noexcept
         = 0;
 
+    /// @var mIoService
     /// Internal IO-service
     boost::asio::io_service mIoService;
 
-    /// Socket
+    /// @var mSocket
+    /// Socket for connections
     boost::asio::ip::tcp::socket mSocket;
 
-    /// Resolver
+    /// @var mResolver
+    /// Endpoint resolver
     boost::asio::ip::tcp::resolver mResolver;
 
+    /// @var mResponse
     /// Response string
     std::string mResponse;
 
+    /// @var mBuffer
     /// Read buffer
     boost::asio::streambuf mBuffer;
 
+    /// @var mHost
     /// Hostname
     const std::string mHost;
 
+    /// @var mPort
     /// Port
     const std::string mPort;
 
+    /// @var mComponent
     /// Component string used for logging
     const std::string mComponent;
 
+    /// @var mrFeed
     /// Handler Feed reference
     feed::Feed& mrFeed;
 
 private:
+    /// @var mConnectTimer
     /// Connection timer
     boost::asio::deadline_timer mConnectTimer;
 };
 
 }  // namespace client
-}  // namespace network
+}  // namespace feed
